Moved common/cli to cli reporter
