+++ src/robotframework_find_unused/reporter/cli/common.py
@@ -1,9 +1,12 @@
+from typing import TYPE_CHECKING
+
 import click
 
-from .const import FileUseType, KeywordData, VariableData
+if TYPE_CHECKING:
+    from robotframework_find_unused.common.const import FileUseType, KeywordData, VariableData
 
 
-def pretty_kw_name(keyword: KeywordData) -> str:
+def pretty_kw_name(keyword: "KeywordData") -> str:
     """
     Format keyword name for output to the user
     """
@@ -18,7 +21,7 @@ def pretty_kw_name(keyword: KeywordData) -> str:
     return name
 
 
-def pretty_file_path(path: str, file_types: set[FileUseType]) -> str:  # noqa: PLR0911
+def pretty_file_path(path: str, file_types: "set[FileUseType]") -> str:  # noqa: PLR0911
     """
     Format file path for output to the user
     """
@@ -44,7 +47,7 @@ def pretty_file_path(path: str, file_types: set[FileUseType]) -> str:  # noqa: P
     raise ValueError(msg)
 
 
-def pretty_variable(var: VariableData) -> str:
+def pretty_variable(var: "VariableData") -> str:
     """
     Format variable for output to the user
     """
